# docs/openapi_current.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MITA Finance API",
    "version": "2.2.0",
...
    "contact": {
      "name": "MITA Finance API Support",
      "email": "api-support@mita.finance",
      "url": "https://docs.mita.finance/support"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "termsOfService": "https://mita.finance/terms"
  },
  "servers": [
    {
      "url": "https://api.mita.finance",
      "description": "Production Server"
    },
    {
      "url": "https://staging-api.mita.finance",
      "description": "Staging Server"  
    },
    {
      "url": "http://localhost:8000",
      "description": "Local Development Server"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/api/financial/installment-evaluate": {
      "post": {
        "summary": "Evaluate Installment Plan Affordability",
        "description": "Evaluate whether a user can afford an installment plan based on their budget and financial profile.\n\n**Features:**\n- Comprehensive affordability analysis\n- Budget impact assessment\n- Risk evaluation\n- Payment schedule calculation\n- Financial recommendation engine",
        "tags": ["Financial Analysis"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InstallmentEvaluationRequest"
              },
              "example": {
                "price": 2500.00,
                "months": 12
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Installment evaluation completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstallmentEvaluationResponse"
                },
                "example": {
                  "success": true,
                  "data": {
                    "affordable": true,
                    "monthly_payment": 208.33,
                    "total_amount": 2500.00,
                    "budget_impact": 15.2,
                    "risk_score": "low",
                    "confidence_score": 0.87,
                    "recommendations": [
                      "Monthly payment fits comfortably within your budget",
                      "Consider setting aside an additional $20/month for unexpected expenses"
                    ]
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
    "/api/financial/dynamic-budget-method": {
      "get": {
        "summary": "Get Personalized Budget Method",
        "description": "Retrieve personalized budget allocation recommendation based on user context instead of hardcoded 50/30/20.\\n\\n**Features:**\\n- Income elasticity theory application\\n- Regional cost adjustments\\n- Behavioral economics principles\\n- Dynamic threshold calculation",
        "tags": ["Financial Analysis", "AI Features"],
        "responses": {
          "200": {
            "description": "Personalized budget method",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "personalized_method": {
                          "type": "object",
                          "description": "Budget allocation percentages"
                        },
                        "user_tier": {
                          "type": "string",
                          "example": "calculated_dynamically"
                        },
                        "economic_justification": {
                          "type": "string",
                          "example": "Budget allocation calculated using income elasticity theory, regional cost adjustments, and behavioral economics principles"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Bearer token authentication. Format: `Bearer <jwt_token>`\n\n**Security Features:**\n- 30 minute expiration for access tokens\n- 90 day expiration for refresh tokens\n- Automatic token refresh capability\n- Token revocation support\n- Rate limiting protection\n\n**Usage:**\n```\nAuthorization: Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9...\n```"
      }
    },
    "schemas": {
...
      "InstallmentEvaluationRequest": {
        "type": "object",
        "required": ["price", "months"],
        "properties": {
          "price": {
            "type": "number",
            "format": "float",
            "minimum": 1.0,
            "maximum": 1000000.0,
            "example": 2500.00,
            "description": "Total price of the item"
          },
          "months": {
            "type": "integer",
            "minimum": 1,
            "maximum": 60,
            "example": 12,
            "description": "Number of installment months"
          }
        }
      },
      "InstallmentEvaluationResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "properties": {
              "affordable": {
                "type": "boolean",
                "example": true,
                "description": "Whether the installment is affordable"
              },
              "monthly_payment": {
                "type": "number",
                "format": "float",
                "example": 208.33,
                "description": "Monthly payment amount"
              },
              "total_amount": {
                "type": "number",
                "format": "float",
                "example": 2500.00
              },
              "budget_impact": {
                "type": "number",
                "format": "float",
                "example": 15.2,
                "description": "Percentage of budget impact"
              },
              "risk_score": {
                "type": "string",
                "enum": ["low", "medium", "high"],
                "example": "low"
              },
              "confidence_score": {
                "type": "number",
                "format": "float",
                "minimum": 0.0,
                "maximum": 1.0,
                "example": 0.87,
                "description": "Analysis confidence score"
              },
              "recommendations": {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "example": [
                  "Monthly payment fits comfortably within your budget",
                  "Consider setting aside an additional $20/month for unexpected expenses"
                ]
              }
            }
          }
        }
      },
...
      "StandardizedErrorResponse": {
        "type": "object",
        "required": ["success", "error"],
        "properties": {
          "success": {
            "type": "boolean",
            "example": false,
            "description": "Always false for error responses"
          },
          "error": {
            "type": "object",
            "required": ["code", "message", "error_id", "timestamp"],
            "properties": {
              "code": {
                "type": "string",
                "example": "VALIDATION_2002",
                "description": "Machine-readable error code for programmatic handling"
              },
              "message": {
                "type": "string",
                "example": "Invalid input data",
                "description": "Human-readable error message"
              },
              "error_id": {
                "type": "string",
                "example": "mita_507f1f77bcf8",
                "description": "Unique error identifier for support and debugging"
              },
              "timestamp": {
                "type": "string",
                "format": "date-time",
                "example": "2024-01-15T10:30:00Z",
                "description": "When the error occurred"
              },
              "details": {
                "type": "object",
                "description": "Additional error context and validation details"
              }
            }
          }
        }
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Unauthorized - Invalid or missing authentication token",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/StandardizedErrorResponse"
            },
            "examples": {
              "missing_token": {
                "summary": "Missing Authorization Header",
                "value": {
                  "success": false,
                  "error": {
                    "code": "AUTH_TOKEN_MISSING",
                    "message": "Authorization header is required",
                    "error_id": "mita_507f1f77bcf8",
                    "timestamp": "2024-01-15T10:30:00Z"
                  }
                }
              },
              "invalid_token": {
                "summary": "Invalid Token",
                "value": {
                  "success": false,
                  "error": {
                    "code": "AUTH_TOKEN_INVALID",
                    "message": "Invalid or malformed token",
                    "error_id": "mita_507f1f77bcf8",
                    "timestamp": "2024-01-15T10:30:00Z"
                  }
                }
              },
              "expired_token": {
                "summary": "Expired Token",
                "value": {
                  "success": false,
                  "error": {
                    "code": "AUTH_TOKEN_EXPIRED",
                    "message": "Access token has expired",
                    "error_id": "mita_507f1f77bcf8",
                    "timestamp": "2024-01-15T10:30:00Z",
                    "details": {
                      "expires_at": "2024-01-15T09:30:00Z",
                      "refresh_required": true
                    }
                  }
                }
              }
            }
          }
        }
      },
      "ValidationError": {
        "description": "Validation Error - Invalid input data",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/StandardizedErrorResponse"
            },
            "examples": {
              "field_validation": {
                "summary": "Field Validation Errors",
                "value": {
                  "success": false,
                  "error": {
                    "code": "VALIDATION_INVALID_FORMAT",
                    "message": "Invalid input data",
                    "error_id": "mita_507f1f77bcf8",
                    "timestamp": "2024-01-15T10:30:00Z",
                    "details": {
                      "validation_errors": [
                        {
                          "field": "email",
                          "message": "Invalid email format",
                          "value": "invalid-email"
                        },
                        {
                          "field": "amount",
                          "message": "Amount must be positive",
                          "value": -50.0
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      },
...
    }
  },
...
}